Upload completed files from lab
Included are 2 python files that take audio input from the user, run it through Watson speech-to-text, return the audio as text, send the text as a prompt to ChatGPT-3, send the response from ChatGPT-3 to Watson text-to-speech and return the response in the form of audio in the voice selected by the user. A Flask app is also set up in Server.py. Also included are the Docker file and its requirements.txt., configuration files for the stt and tts models, and an interface already provided by IBM as building the interface is not within the scope of this lab.